extracted all slice definitions from store.js into dedicated files under src/store
