+++ src/components/Header.js
@@ -1,6 +1,6 @@
 import classes from "./Header.module.css";
 import { useSelector, useDispatch } from "react-redux";
-import { authActions } from "../store";
+import { authActions } from "../store/auth-slice";
 
 const Header = () => {
   const dispatch = useDispatch();
